% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_text.R
\name{col_to_text}
\alias{col_to_text}
\title{Dataframe Column to Text File}
\usage{
col_to_text(dat, text_col = "document_text", dest_dir)
}
\arguments{
\item{dat}{Dataframe to process}

\item{text_col}{Column to save as text files}

\item{dest_dir}{Directory to save output}
}
\value{
Text files for
}
\description{
BTRIS allows queries and downloads of free-text notes in CRIS as a CSV file.
This function takes a pre-processed dataframe of this data and saves each note as a separate text file.
Filenames are formatted as such: 'mrn,document_date,document_name(note_index).txt'
}
