% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{elapsed_years}
\alias{elapsed_years}
\title{Calculate number of years between two dates}
\usage{
elapsed_years(start_date, end_date)
}
\arguments{
\item{start_date}{Date}

\item{end_date}{Date}
}
\value{
dbl
}
\description{
\href{https://stackoverflow.com/questions/15569333/get-date-difference-in-years-floating-point}{Stack Overflow}
}
\examples{
elapsed_years(lubridate::mdy("01-01-2020"), lubridate::mdy("07-16-2000"))
}
