% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{enrich_set}
\alias{enrich_set}
\title{Enrich an adjacency matrix by linking pairs if they share a given annotation}
\usage{
enrich_set(adj_matrix, term_pair_df)
}
\arguments{
\item{adj_matrix}{A HPO adjacency matrix, see \code{get_adjacency_matrix()}.}

\item{term_pair_df}{Two column dataframe: 'hpo_id' & 'diseases' (a character list of disease annotations). Will make more generic.}
}
\value{
An enriched adjacency matrix.
}
\description{
A clunky way to draw new edges between unlinked terms if they may be linked by a given annotation. E.g. if two HPO terms have a shared Orphanet annotation, draw a new edge.
}
