% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fct_case_when}
\alias{fct_case_when}
\title{dplyr::case_when() that returns factor}
\usage{
fct_case_when(...)
}
\description{
Thanks to https://stackoverflow.com/questions/49572416/r-convert-to-factor-with-order-of-levels-same-with-case-when
}
