% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_adjacency_matrix}
\alias{get_adjacency_matrix}
\title{Get a directed adjacency matrix from \code{ontologyIndex} object}
\usage{
get_adjacency_matrix(ontology, terms)
}
\arguments{
\item{ontology}{\code{ontologyIndex} object}

\item{terms}{named ontology term list}
}
\value{
sqaure directed adjacency matrix
}
\description{
Adapted from package \code{ontologyPlot} so entire package doesn't have to be installed, as it requires Rgraphviz.
}
\examples{
get_adjacency_matrix(hpo, ontologyIndex::get_descendants(hpo, "HP:0000118"))

}
