% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_categories}
\alias{get_categories}
\title{Get HPO categories for a given HPO ID.}
\usage{
get_categories(hpo, term_id)
}
\arguments{
\item{hpo}{\code{ontologyIndex} hpo object (also loaded with \code{PhenoClustR})}

\item{term_id}{single HPO ID}
}
\value{
Named character list of HPO IDs
}
\description{
HPO Categories are defined as children terms of "Phenotypic Abnormality", "HP:0000118". E.g. "Abnormality of Connective Tissue". If no applicable category, returns HPO root HP:0000001.
}
\examples{
get_categories(hpo, "HP:0001025") # Urticaria
}
