% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_undirected_matrix}
\alias{get_undirected_matrix}
\title{Get an undirected adjacency matrix from \code{ontologyIndex} object}
\usage{
get_undirected_matrix(ontology, terms)
}
\arguments{
\item{ontology}{\code{ontologyIndex} object}

\item{terms}{named ontology term list}
}
\value{
sqaure undirected adjacency matrix
}
\description{
Difers from \code{get_adjacency_matrix()} in that it returns an undirected matrix. Good starting point for enrichment.
}
\examples{
get_undirected_matrix(hpo, ontologyIndex::get_descendants(hpo, "HP:0000118"))

}
