% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_umap.R
\name{iterate_umap}
\alias{iterate_umap}
\title{Run UMAP with many combinations of hyperparameters.}
\usage{
iterate_umap(dat, dat.labels, params)
}
\arguments{
\item{dat}{Unlabeled dataframe for UMAP input}

\item{dat.labels}{Labels for dat}

\item{params}{Sqaure dataframe of UMAP hyperparameter values (all required):
\itemize{
\item n_neighbors
\item min_dist
\item n_components
\item metric
\item method
\item seed
}}
}
\value{
Nested tibble::tibble() containing UMAP objs, output layout, labeled output, and 2D/3D plot obj.
}
\description{
Run UMAP with many combinations of hyperparameters.
}
