% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_p.R
\name{pairwise_p}
\alias{pairwise_p}
\title{Pairwise P Values from Similarity Matrix}
\usage{
pairwise_p(sim_matx)
}
\arguments{
\item{sim_matx}{Similarity matrix, output from ontologySimilarity::get_sim_grid()}
}
\value{
Matrix of P values
}
\description{
Pairwise P Values from Similarity Matrix
}
