% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_p.R
\name{pairwise_p_ontology}
\alias{pairwise_p_ontology}
\title{Pairwise P-Values Calculated from Ontology & Term Sets}
\usage{
pairwise_p_ontology(term_set, ontology, ic, method)
}
\arguments{
\item{term_set}{list of character vectors contain ontology term IDs}

\item{ontology}{\code{ontology_index} object}

\item{ic}{information content, named numeric list}

\item{method}{"lin" or "resnik}
}
\value{
matrix of p-values for each pair in term set (by index)
}
\description{
Pairwise P-Values Calculated from Ontology & Term Sets
}
