% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Quick 2D/3D Plot from UMAP Output}
\usage{
plot_umap(umap_obj, label_df, interact = TRUE)
}
\arguments{
\item{umap_obj}{Object returned from \code{umap::umap()}}

\item{label_df}{Dataframe with labels for umap obj layout}

\item{interact}{Use plotly to create interactive plot}
}
\value{
ggplot2 / plotly obj.
}
\description{
Currently only supports dataframe with 'mrn' column for labeling. Will make more general.
}
