% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hpo.R
\name{update_hpo}
\alias{update_hpo}
\title{Update Version of Human Phenotype Ontology}
\usage{
update_hpo()
}
\description{
Internal function to fetch latest version of the Human Phenotype Ontology and save it as an R object, \code{hpo}, availble upon package load.
}
\examples{
update_hpo()
}
